changed return type to float

+++ SubtractClass.java
@@ -1,5 +1,5 @@
 public SubtractClass {
-	public int modular(int x, int y) {
+	public float modular(int x, int y) {
 		return x - y;
 	}
 }
